speed_layer/requirements.txt
pyspark==3.5.6
happybase==1.2.0
thrift==0.22.0
kafka-python==2.0.2
requests==2.32.3
prometheus-client==0.20.0